REFACTOR: upload policy
