+++ sample/src/main/java/com/imagekit/android/sample/MainActivity.kt
@@ -17,7 +17,7 @@ class MainActivity : AppCompatActivity() {
             publicKey = "public_5P5QM23aRv9XkOcfJO1okZ0DzOw=",
             urlEndpoint = "https://ik.imagekit.io/tqhfz73me",
             transformationPosition = TransformationPosition.PATH,
-            authenticationEndpoint = "https://bd1e-2401-4900-1c54-e034-90c7-c25b-fbd1-1769.ngrok-free.app/"
+            authenticationEndpoint = "https://027e-119-82-70-86.ngrok-free.app/"
         )
 
         btUploadImage.setOnClickListener {
